std_bitset.h (_M_do_find_next): Fix -Wall nit.
	* include/std/std_bitset.h (_M_do_find_next): Fix -Wall nit.
	* include/bits/concept_check.h: Fix multi-line comment.
	* testsuite/17_intro/headers.cc (dg-options): Add -Wall -Wsystem-header
	when target is *-*-freebsd*.

From-SVN: r65620

+++ libstdc++-v3/include/std/std_bitset.h
@@ -414,7 +414,7 @@ namespace std
       _M_do_find_next(size_t __prev, size_t __not_found) const
       {
 	++__prev;
-	if (__prev >= _GLIBCPP_BITSET_BITS_PER_WORD)
+	if (__prev >= ((size_t) _GLIBCPP_BITSET_BITS_PER_WORD))
 	  return __not_found;
 
 	_WordT __x = _M_w >> __prev;
